:recycle: Split tests of destruction endpoints
